Move settings widgets back
