+++ lib/view/common/yaru_single_info_row.dart
