+++ lib/view/common/yaru_slider_row.dart
