+++ lib/view/pages/mouse_and_touchpad/general_section.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:settings/constants.dart';
+import 'package:settings/view/common/yaru_toggle_buttons_row.dart';
 import 'package:settings/view/settings_section.dart';
-import 'package:yaru_settings/yaru_settings.dart';
 
 import 'mouse_and_touchpad_model.dart';
