+++ lib/view/pages/privacy/location_page.dart
@@ -3,12 +3,12 @@ import 'package:provider/provider.dart';
 import 'package:settings/constants.dart';
 import 'package:settings/l10n/l10n.dart';
 import 'package:settings/services/settings_service.dart';
+import 'package:settings/view/common/yaru_switch_row.dart';
 import 'package:settings/view/link.dart';
 import 'package:settings/view/pages/privacy/location_model.dart';
 import 'package:settings/view/pages/settings_page.dart';
 import 'package:settings/view/section_description.dart';
 import 'package:settings/view/settings_section.dart';
-import 'package:yaru_settings/yaru_settings.dart';
 
 const kPrivacyUrl = 'https://location.services.mozilla.com/privacy';
