+++ lib/view/pages/sound/sound_page.dart
@@ -3,10 +3,10 @@ import 'package:provider/provider.dart';
 import 'package:settings/constants.dart';
 import 'package:settings/l10n/l10n.dart';
 import 'package:settings/services/settings_service.dart';
+import 'package:settings/view/common/yaru_switch_row.dart';
 import 'package:settings/view/pages/settings_page.dart';
 import 'package:settings/view/pages/sound/sound_model.dart';
 import 'package:settings/view/settings_section.dart';
-import 'package:yaru_settings/yaru_settings.dart';
 
 class SoundPage extends StatelessWidget {
   const SoundPage({Key? key}) : super(key: key);
