--- packages/yaru_settings/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: eb6d86ee27deecba4a83536aa20f366a6044895c
-  channel: stable
-
-project_type: package
