--- packages/yaru_settings/lib/yaru_settings.dart
@@ -1,8 +1,0 @@
-library yaru_settings;
-
-export 'src/yaru_checkbox_row.dart';
-export 'src/yaru_extra_option_row.dart';
-export 'src/yaru_single_info_row.dart';
-export 'src/yaru_slider_row.dart';
-export 'src/yaru_switch_row.dart';
-export 'src/yaru_toggle_buttons_row.dart';
