--- packages/yaru_settings/pubspec.yaml
@@ -1,25 +1,0 @@
-name: yaru_settings
-publish_to: "none"
-
-environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.3.0"
-
-dependencies:
-  flutter:
-    sdk: flutter
-  yaru: ^0.9.0
-  yaru_widgets: ^2.6.0
-
-dev_dependencies:
-  flutter_lints: ^1.0.0
-  flutter_test:
-    sdk: flutter
-  test: ^1.21.4
-
-flutter:
-  uses-material-design: true
-
-
-dependency_overrides:
-  yaru_icons: ^2.1.0
