+++ pubspec.yaml
@@ -47,8 +47,6 @@ dependencies:
   yaru: ^0.9.0
   yaru_colors: ^0.1.6
   yaru_icons: ^2.1.0
-  yaru_settings:
-    path: packages/yaru_settings
   yaru_widgets: ^2.6.0
 
 dev_dependencies:
