+++ test/widgets/yaru_check_box_row_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:yaru_settings/src/yaru_checkbox_row.dart';
+import 'package:settings/view/common/yaru_checkbox_row.dart';
 
 void main() {
   testWidgets('YaruCheckboxRow Test', (WidgetTester tester) async {
