+++ test/widgets/yaru_switch_row_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:settings/view/common/yaru_switch_row.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
-import 'package:yaru_settings/src/yaru_switch_row.dart';
 
 void main() {
   testWidgets('- YaruSwitchRow Test', (WidgetTester tester) async {
