+++ test/widgets/yaru_togggle_button_row_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:yaru_settings/src/yaru_toggle_buttons_row.dart';
+import 'package:settings/view/common/yaru_toggle_buttons_row.dart';
 
 void main() {
   testWidgets('- YaruToggleButtonsRow Test', (WidgetTester tester) async {
